Clean up code in DataConfirm, assigned functions to new files
