+++ src/frontend/project-src/utils/DeepSeekToTable.jsx
@@ -1,0 +1,27 @@
+function convertDeepSeekToTable(parsedData) {
+  if (!Array.isArray(parsedData) || parsedData.length === 0) {
+    return {
+      table: {
+        rowHeaderTitle: "Row",
+        columns: [],
+        rows: [],
+      },
+    };
+  }
+
+  const columns = Object.keys(parsedData[0]);
+  const rows = parsedData.map((rowObj, idx) => ({
+    header: `Row ${idx + 1}`,
+    cells: columns.map((col) => rowObj[col] ?? ""),
+  }));
+
+  return {
+    table: {
+      rowHeaderTitle: "Row",
+      columns,
+      rows,
+    },
+  };
+}
+
+export default convertDeepSeekToTable;
