+++ src/frontend/project-src/utils/TableToDeepSeek.jsx
@@ -1,0 +1,13 @@
+function convertTableToDeepSeekFormat(table) {
+  const { columns, rows } = table;
+
+  return rows.map((row) => {
+    const obj = {};
+    columns.forEach((col, i) => {
+      obj[col] = row.cells[i] ?? "";
+    });
+    return obj;
+  });
+}
+
+export default convertTableToDeepSeekFormat;
